Refactor utils into sub package
utils/collector.py
utils/output.py
utils/filetype.py
utils/validations.py
utils/xz.py
