+++ azurectl/compute_endpoint_task.py
@@ -61,8 +61,8 @@ options:
 # project
 from cli_task import CliTask
 from azure_account import AzureAccount
-from data_collector import DataCollector
-from data_output import DataOutput
+from utils.collector import DataCollector
+from utils.output import DataOutput
 from endpoint import Endpoint
 from help import Help
