+++ azurectl/compute_request_task.py
@@ -32,8 +32,8 @@ options:
 # project
 from cli_task import CliTask
 from azure_account import AzureAccount
-from data_collector import DataCollector
-from data_output import DataOutput
+from utils.collector import DataCollector
+from utils.output import DataOutput
 from request_result import RequestResult
 from help import Help
