+++ azurectl/compute_reserved_ip_task.py
@@ -37,8 +37,8 @@ options:
 # project
 from cli_task import CliTask
 from azure_account import AzureAccount
-from data_collector import DataCollector
-from data_output import DataOutput
+from utils.collector import DataCollector
+from utils.output import DataOutput
 from reserved_ip import ReservedIp
 from help import Help
