+++ azurectl/compute_storage_task.py
@@ -94,8 +94,8 @@ from pytz import utc
 # project
 from cli_task import CliTask
 from azure_account import AzureAccount
-from data_collector import DataCollector
-from data_output import DataOutput
+from utils.collector import DataCollector
+from utils.output import DataOutput
 from logger import log
 from azurectl_exceptions import AzureInvalidCommand
 from storage import Storage
