+++ azurectl/utils/collector.py
