+++ test/unit/utils_filetype_test.py
@@ -1,6 +1,6 @@
 
 
-from azurectl.filetype import FileType
+from azurectl.utils.filetype import FileType
 
 from azurectl.azurectl_exceptions import *
